refactor: promote Aspect to top-level enum, delete IncrementalBuild
Move the Aspect enum from IncrementalBuild.Aspect to a top-level type
and inline the one-liner isEmptyOrIgnorable() at its single call site.
Delete IncrementalBuild.java — it contained no other logic after the
BuildContext migration.
